MAINT conversion from dependency to constituency gets its own submodule
