+++ educe/rst_dt/tests.py
@@ -6,6 +6,7 @@ import unittest
 import copy
 
 from educe.rst_dt import annotation, parse, SimpleRSTTree
+from educe.rst_dt.dep2con import deptree_to_simple_rst_tree
 from educe.rst_dt.deptree import RstDepTree
 from educe.rst_dt.parse import (parse_lightweight_tree,
                                 parse_rst_dt_tree,
@@ -135,13 +136,13 @@ class RSTTest(unittest.TestCase):
         for lstr in lw_trees:
             rst1 = parse_lightweight_tree(lstr)
             dep = RstDepTree.from_simple_rst_tree(rst1)
-            rst2 = dep.to_simple_rst_tree([])
+            rst2 = deptree_to_simple_rst_tree(dep, [])
             self.assertEqual(rst1, rst2, "round-trip on " + lstr)
 
         for name, tree in self._test_trees().items():
             rst1 = SimpleRSTTree.from_rst_tree(tree)
             dep = RstDepTree.from_simple_rst_tree(rst1)
-            rst2 = dep.to_simple_rst_tree([])
+            rst2 = deptree_to_simple_rst_tree(dep, [])
             self.assertEqual(treenode(rst1).span,
                              treenode(rst2).span,
                              "span equality on " + name)
@@ -178,16 +179,16 @@ class RSTTest(unittest.TestCase):
             dep = RstDepTree.from_simple_rst_tree(rst1)
 
             dep_a = dep
-            rst2a = dep_a.to_simple_rst_tree([])
+            rst2a = deptree_to_simple_rst_tree(dep_a, [])
             self.assertEqual(rst1, rst2a, "round-trip on " + lstr)
 
             dep_b = copy.deepcopy(dep)
             dep_b.deps[0].reverse()
-            rst2b = dep_b.to_simple_rst_tree([])
+            rst2b = deptree_to_simple_rst_tree(dep_b, [])
 
             dep_c = copy.deepcopy(dep)
             random.shuffle(dep_c.deps[0])
-            rst2c = dep_c.to_simple_rst_tree([])
+            rst2c = deptree_to_simple_rst_tree(dep_c, [])
 
     def test_rst_to_dt_nuclearity_loss(self):
         """
@@ -221,10 +222,10 @@ class RSTTest(unittest.TestCase):
 
         # a little sanity check first
         dep0 = RstDepTree.from_simple_rst_tree(rst0)
-        rev0 = dep0.to_simple_rst_tree(['r'])
+        rev0 = deptree_to_simple_rst_tree(dep0, ['r'])
         self.assertEqual(rst0, rev0, "same structure " + nuked)  # sanity
 
         # now the real test
         dep1 = RstDepTree.from_simple_rst_tree(rst1)
-        rev1 = dep1.to_simple_rst_tree(['r'])
+        rev1 = deptree_to_simple_rst_tree(dep1, ['r'])
         # self.assertEqual(rst0, rev1, "same structure " + tricky)
