+++ educe/rst_dt/util/cmd/deptree.py
@@ -56,7 +56,7 @@ def convert(corpus, multinuclear, odir):
         with open(os.path.join(dt_dir, suffix), 'w') as fout:
             fout.write(str(dtree))
 
-        stree2 = dtree.to_simple_rst_tree(multinuclear)
+        stree2 = deptree_to_simple_rst_tree(dtree, multinuclear)
         with open(os.path.join(rst2_dir, suffix), 'w') as fout:
             fout.write(str(stree2))
